refactor(gs): extract CertificateCell into its own component
Moves the CertificateCell render helper and its expiry/severity/ready
helpers and styles out of DeploymentHttpRoutesCard into a dedicated
CertificateCell/ directory, per the repo's component-structure
convention.
