+++ plugins/gs/src/components/deployments/deployment-details/DeploymentGateway/DeploymentHttpRoutesCard/CertificateCell/index.ts
@@ -1,0 +1,1 @@
+export { CertificateCell } from './CertificateCell';
